move MyAccount component to Setting component
