+++ client/src/components/Pages/Settings/Intern/index.js
@@ -5,7 +5,7 @@ import * as T from '../../../Common/Typography';
 import { SETTINGS } from '../../../../constants/navRoutes';
 import * as S from './style';
 
-import MyAccount from './MyAccount';
+import MyAccount from '../MyAccount';
 import MyProfile from './MyProfile';
 import AboutMe from './AboutMe';
 import Verifications from './Verifications';
