+++ client/src/components/Pages/Settings/MyAccount.js
@@ -1,16 +1,16 @@
 import React, { useState } from 'react';
 import axios from 'axios';
 import { CloseOutlined } from '@ant-design/icons';
-import { Input } from '../../../Common/Inputs';
-import { Col, Row } from '../../../Common/Grid';
-import * as S from './style';
-import * as T from '../../../Common/Typography';
-import Button from '../../../Common/ButtonNew';
-import { API_INTERN_SETTINGS_MY_ACCOUNT } from '../../../../constants/apiRoutes';
-import Notification from '../../../Common/Notification';
-import { SETTINGS } from '../../../../constants/navRoutes';
+import { Input } from '../../Common/Inputs';
+import { Col, Row } from '../../Common/Grid';
+import * as S from './Host/style';
+import * as T from '../../Common/Typography';
+import Button from '../../Common/ButtonNew';
+import { API_INTERN_SETTINGS_MY_ACCOUNT } from '../../../constants/apiRoutes';
+import Notification from '../../Common/Notification';
+import { SETTINGS } from '../../../constants/navRoutes';
 
-const { validate, internSettings } = require('../../../../validation');
+const { validate, internSettings } = require('../../../validation');
 
 const MyAccount = props => {
   const [state, setState] = useState({
